split membership from member.go
